+++ src/vars/mtxdb/mtx_free.c
@@ -6,7 +6,7 @@
 /*   By: topiana- <topiana-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/26 16:13:50 by topiana-          #+#    #+#             */
-/*   Updated: 2025/04/01 16:45:38 by topiana-         ###   ########.fr       */
+/*   Updated: 2025/04/01 22:37:25 by topiana-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@ void	mtx_free(char **mtx);
 /* Default free mtx-pointer */
 void	mtx_free(char **mtx)
 {
-	unsigned int	i;
+	int	i;
 
 	if (mtx == NULL)
 		return ;
 	i = 1;
-	while (i < (unsigned int)*(int *)mtx)
+	while (i < *(int *)mtx)
 	{
 		free(mtx[i]);
 		i++;
